requirements.txt
selenium==4.3
urllib
pandas
time
unittest
glob
os
requests
unicodedata
arrow
bs4
numpy
datetime
sklearn.preprocessing
webdriver_manager.chrome
chromedriver-binary==77.0.3865.40.0
gunicorn==19.9.0
uvicorn